+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 4.0.1.pre
 ---------------
+ * Create index m170922_023840__alter_table__cms_content_element_property
  * Update console cms/utils/clear-all-thumbnails
  * Fixed admin
  * New console cmd for delete content elements